misc: move pdf, sgpu, and wgsl_printer out of experimental
