+++ tests/interpreters/test_wgsl_printer.py
@@ -3,7 +3,7 @@ from io import StringIO
 from xdsl.context import MLContext
 from xdsl.dialects import arith, builtin, gpu, memref, test
 from xdsl.dialects.builtin import IndexType, IntegerAttr, IntegerType, i32
-from xdsl.interpreters.experimental.wgsl_printer import WGSLPrinter
+from xdsl.interpreters.wgsl_printer import WGSLPrinter
 from xdsl.parser import Parser
 from xdsl.utils.test_value import TestSSAValue
