+++ xdsl/interpreters/wgpu.py
@@ -9,8 +9,8 @@ from xdsl.dialects import gpu
 from xdsl.dialects.builtin import IndexType
 from xdsl.dialects.memref import MemRefType
 from xdsl.interpreter import Interpreter, InterpreterFunctions, impl, register_impls
-from xdsl.interpreters.experimental.wgsl_printer import WGSLPrinter
 from xdsl.interpreters.shaped_array import ShapedArray
+from xdsl.interpreters.wgsl_printer import WGSLPrinter
 from xdsl.ir import Attribute, SSAValue
 from xdsl.traits import SymbolTable
 from xdsl.utils.hints import isa
